Add nsfallback feature for browser request, closes #2261
nsfallback is a fallback ns search when requesting:
foo:bar -> foo:bar:$conf['start'] -> foo:bar:bar

nsfallback is similar to autoplural, but autoplural is only effective during internal link parsing (not effective during browser request i.e. getID()). Not sure why it works like that, but because it's not rewriting getID you can try creating page by visiting the ID directly - however it is not working by searching for the not existed page or linking it, which is the recommended way in https://www.dokuwiki.org/page#create_a_page.

I decided to implement nsfallback to apply only during browser request and $ACT=show, not during internal link parsing, because this is most useful when users visit /foo/bar looking for /foo/bar/(start) - this is less intended for people writing the link on the wiki. By not rewriting the internal link, user can still create pages by linking it or searching.

This option is off by default. Tests for getid and resolve_pageid have been added to define the behavior.